Refactored articles routes
