+++ src/containers/Articles/ArticlesContainer.scss
@@ -1,4 +1,4 @@
-.ArticlesLayout {
+.ArticlesContainer {
   height: 100%;
   padding-top: 24px;
 }
