+++ src/containers/index.js
@@ -1,18 +1,24 @@
 export App from './App/App';
-export ArticlesPage from './ArticlesPage/ArticlesPage';
-export ArticlePage from './ArticlePage/ArticlePage';
+export NotFound from './NotFound/NotFound';
+// LAYOUT
+export MainFooter from './MainFooter/MainFooter';
+export MainNavbar from './MainNavbar/MainNavbar';
+// ARTCILES
+export ArticlesContainer from './Articles/ArticlesContainer';
+export ArticlesPage from './Articles/ArticlesPage/ArticlesPage';
+export ArticlePage from './Articles/ArticlePage/ArticlePage';
+// LOGIN
+// TODO: this needs to be cleaned-up
+export Login from './Login/Login';
+export LoginDialog from './LoginDialog/LoginDialog';
+export LoginSuccess from './LoginSuccess/LoginSuccess';
+
 export BestPracticesPage from './BestPracticesPage/BestPracticesPage';
 export Chat from './Chat/Chat';
 export Conversation from './Conversation/Conversation';
 export DialogsContainer from './DialogsContainer/DialogsContainer';
 export HomePage from './HomePage/HomePage';
-export Login from './Login/Login';
-export LoginDialog from './LoginDialog/LoginDialog';
-export LoginSuccess from './LoginSuccess/LoginSuccess';
-export MainFooter from './MainFooter/MainFooter';
-export MainNavbar from './MainNavbar/MainNavbar';
 export NewArticlePage from './NewArticlePage/NewArticlePage';
-export NotFound from './NotFound/NotFound';
 export TutorialsPage from './TutorialsPage/TutorialsPage';
 export ProfilePage from './ProfilePage/ProfilePage';
 export UsersPage from './UsersPage/UsersPage';
